[oss-timeseries] move runs/ to runs_legacy/
