+++ tensorboard/webapp/plugins/npmi/views/data_selection/metric_search/BUILD
@@ -58,7 +58,7 @@ tf_ts_library(
         "//tensorboard/webapp/plugins/npmi/store",
         "//tensorboard/webapp/plugins/npmi/store:types",
         "//tensorboard/webapp/plugins/npmi/testing",
-        "//tensorboard/webapp/runs",
+        "//tensorboard/webapp/runs_legacy",
         "//tensorboard/webapp/testing:mat_icon",
         "@npm//@angular/common",
         "@npm//@angular/core",
