+++ tensorboard/webapp/plugins/npmi/views/main/BUILD
@@ -34,7 +34,7 @@ ng_module(
         "//tensorboard/webapp/plugins/npmi/views/data_selection",
         "//tensorboard/webapp/plugins/npmi/views/selected_annotations",
         "//tensorboard/webapp/plugins/npmi/views/violin_filters",
-        "//tensorboard/webapp/runs",
+        "//tensorboard/webapp/runs_legacy",
         "@npm//@angular/core",
         "@npm//@angular/forms",
         "@npm//@ngrx/store",
@@ -58,7 +58,7 @@ tf_ts_library(
         "//tensorboard/webapp/plugins/npmi/actions",
         "//tensorboard/webapp/plugins/npmi/store",
         "//tensorboard/webapp/plugins/npmi/testing",
-        "//tensorboard/webapp/runs",
+        "//tensorboard/webapp/runs_legacy",
         "@npm//@angular/core",
         "@npm//@angular/platform-browser",
         "@npm//@ngrx/store",
